requirements.txt
# Streamlit App Framework
streamlit==1.35.0

# Resume Parsing
pdfplumber==0.10.2
docx2txt==1.0

# NLP and Skill Extraction
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Embeddings & Matching
sentence-transformers==2.6.1
scikit-learn==1.4.2
numpy==1.26.4
scipy==1.13.0

# Similarity & String Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Visualization & Debugging
matplotlib==3.9.0
pandas==2.2.2

# Optional: if you plan to use OpenAI/GPT for resume tailoring
openai==1.25.2
tiktoken==0.6.0

# Optional: progress bars, UX sugar
tqdm==4.66.2
